+++ examples/ecs_cronjob/infra/provider.tf
@@ -1,5 +1,8 @@
 provider "aws" {
+  region = "eu-central-1"
+
   # Make it faster by skipping something
+  skip_get_ec2_platforms      = true
   skip_metadata_api_check     = true
   skip_region_validation      = true
   skip_credentials_validation = true
@@ -9,9 +12,9 @@ provider "aws" {
 
   default_tags {
     tags = {
-      Project     = "terra3"
-      Environment = "qa"
-      Terraform   = "true"
+      "terra3.io/project" = "terra3-examples"
+      Environment         = "qa"
+      Terraform           = "true"
     }
   }
 }